AK: Moved TypedTransfer into it's own header.
